+++ src/main/java/com/skills/skills/controllers/AuthenticationController.java
@@ -4,9 +4,13 @@ import com.skills.skills.data.SkillsCategoryRepository;
 import com.skills.skills.data.SkillsRepository;
 import com.skills.skills.data.TagRepository;
 import com.skills.skills.data.UserRepository;
+import com.skills.skills.models.user.User;
+import com.skills.skills.models.user.UserProfile;
 import com.skills.skills.models.*;
 import com.skills.skills.models.dto.LoginFormDTO;
 import com.skills.skills.models.dto.RegisterFormDTO;
+import com.skills.skills.models.user.User;
+import com.skills.skills.models.user.UserProfile;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -58,6 +62,13 @@ public class AuthenticationController {
         session.setAttribute(userSessionKey, user.getId());
 
     }
+//
+//    @GetMapping
+//    public String displayPageAfterLogin(HttpSession session, Model model) {
+//        User user = getUserFormSession(session);
+//        model.addAttribute("user", user);
+//        return "index";
+//    }
 
     @GetMapping("/users/profile")
     public String displayPageAfterLogin (HttpSession session, Model model) {
