+++ src/main/java/com/skills/skills/data/SkillsCategoryRepository.java
@@ -1,7 +1,7 @@
 package com.skills.skills.data;
 
-import com.skills.skills.models.SkillsCategory;
-import org.springframework.data.repository.PagingAndSortingRepository;
+import com.skills.skills.models.skill.SkillsCategory;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
