+++ src/main/java/com/skills/skills/data/SkillsRepository.java
@@ -1,7 +1,7 @@
 package com.skills.skills.data;
 
-import com.skills.skills.models.Skill;
-import org.springframework.data.repository.PagingAndSortingRepository;
+import com.skills.skills.models.skill.Skill;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
