+++ src/main/java/com/skills/skills/data/UserRepository.java
@@ -1,7 +1,6 @@
 package com.skills.skills.data;
 
-import com.skills.skills.models.Skill;
-import com.skills.skills.models.User;
+import com.skills.skills.models.user.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
