--- src/main/java/com/skills/skills/models/EventDetails.java
@@ -1,4 +1,0 @@
-package com.skills.skills.models;
-
-public class EventDetails {
-}
